"int to Integer" & "double to Double" changes Eingangsrechnung Mitarbeiter ZeitErfassung DALDatabase IDAL
git-svn-id: file:///home/node/svn/project_se/goran/EPU_Office@143 bde5ac68-c2e4-480a-bd7c-0859b87ef666

+++ src/model/Mitarbeiter.java
@@ -14,7 +14,7 @@ public class Mitarbeiter extends AbstractObject {
 
     private String vorname;
     private String nachname;
-    private int stundenSatz;
+    private Double stundenSatz;
     private Date geburtsDatum;
 
     public Date getGeburtsDatum() {
@@ -33,11 +33,11 @@ public class Mitarbeiter extends AbstractObject {
         this.nachname = nachname;
     }
 
-    public int getStundenSatz() {
+    public Double  getStundenSatz() {
         return stundenSatz;
     }
 
-    public void setStundenSatz(int stundenSatz) {
+    public void setStundenSatz(Double stundenSatz) {
         this.stundenSatz = stundenSatz;
     }
